server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary
alembic

# Pydantic (data validation, included with FastAPI but good to be explicit)
pydantic
pydantic-settings

# LangChain and LangGraph
langchain
langgraph
langchain-groq
langchain-google-genai
python-dotenv

# For asynchronous HTTP requests (used by LangChain)
httpx

alembic

# Vector Database and Embedding API
chromadb
google-generativeai

# ... (keep all existing lines)

# For ML Model Training
pandas
scikit-learn